// src/models/vw_dw_reportingview_offenses.ts
export enum VWDWReportingViewOffensesColumns {
  offense_name = 'offense_name',
  offense_category_name = 'offense_category_name',
  case_type = 'case_type',
  court_name = 'court_name',
  statute_name = 'statute_name',
  arrest_id = 'arrest_id',
  create_time = 'create_time',
  local_id = 'local_id',
  age = 'age',
  age_group_seq = 'age_group_seq',
  age_group_name = 'age_group_name',
  name_first = 'name_first',
  name_middle = 'name_middle',
  name_last = 'name_last',
  name_suffix = 'name_suffix',
  name_full = 'name_full',
  gender_code = 'gender_code',
  gender_name = 'gender_name',
  location_city = 'location_city',
  location_county = 'location_county',
  location_state_code = 'location_state_code',
  location_state = 'location_state',
  arrest_date = 'arrest_date',
  arrest_day_of_week = 'arrest_day_of_week',
  arrest_day_of_week_name = 'arrest_day_of_week_name',
  arrest_date_hour = 'arrest_date_hour',
  arrest_date_hour_index = 'arrest_date_hour_index',
  release_day_of_week = 'release_day_of_week',
  release_day_of_week_name = 'release_day_of_week_name',
  release_date_hour = 'release_date_hour',
  release_date_hour_index = 'release_date_hour_index',
  arrest_datetime = 'arrest_datetime',
  release_date = 'release_date',
  release_datetime = 'release_datetime',
  days_in = 'days_in',
  charges_count = 'charges_count',
  total_bail = 'total_bail',
}